Changed the solver to arma::spsolve
This needed the implementation of a getSparseStorage() method in the BaseMatrix class.
This is not ideal as the matrix is converted from dense to sparse at every iteration. (loop over the whole matrix).
Regardless, the code runs faster, in the future, consider storing the stifnesses mateices in sparse matrices from the start.